+++ packages/ub-pub/locale/lang-en.js
@@ -170,10 +170,4 @@ UB.i18nExtend({
   fieldValidationError: 'Cannot save <b>{0}</b> card.<br > Invalid field',
   blankText: 'This field is required',
   licenseExceedText: 'Application license limits is exceed. You can view data but most of modification is not allowed'
-
-  aclRlsPane: {
-    title: 'Access',
-    inclusive: 'Inclusive',
-    remove: 'Remove'
-  }
 })